requirements.txt
PyQt5 == 5.15.11
mysql-connector-python == 9.1.0
pandas == 2.2.3
pymysql
